CoreDataViewModel: - Add deleteData function
Work+CoreDataProperties:
- Change overTime to double
- Add workTime
- Implement the change in rest of classes referencing work

SettingsView:
- Add controls for maximum overtime allowed
- Add section for user data control
- Remove seconds from timer pickers

+++ ClockIn/View/HistoryRow.swift
@@ -17,7 +17,7 @@ struct HistoryRow: View {
     var date: Date
     var startDate: Date
     var finishDate: Date
-    var overTime: Float
+    var overTime: Double
     
     @State var isShowingDetails: Bool = false 
     @State var detailType: DetailDisplayType
